Update calculator.py to accept float inputs; change file modes for multiple scripts

+++ cell04/ex02/calculator.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
-a = float(input("Give me the first number: "))
-b = float(input("Give me the second number: "))
+a = int(input("Give me the first number: "))
+b = int(input("Give me the second number: "))
 print("Thank you!")
 print(f"{a} + {b} = {a+b}")
 print(f"{a} - {b} = {a-b}")
